// src/app/model/block.parser.ts
type Nullable<T> = T | null;
type $$RuleType<T> = () => Nullable<T>;
export interface ASTNodeIntf {
    kind: ASTKinds;
}
export enum ASTKinds {
    BLOCK = "BLOCK",
    BLOCK_$0 = "BLOCK_$0",
    BLOCK_CONTENT_1 = "BLOCK_CONTENT_1",
    BLOCK_CONTENT_2 = "BLOCK_CONTENT_2",
    BLOCK_CONTENT_$0 = "BLOCK_CONTENT_$0",
    BLOCK_CONTENT_$1 = "BLOCK_CONTENT_$1",
    NOTE = "NOTE",
    NOTE_$0 = "NOTE_$0",
    SIMPLE_NOTE_1 = "SIMPLE_NOTE_1",
    SIMPLE_NOTE_2 = "SIMPLE_NOTE_2",
    SIMPLE_NOTE_$0 = "SIMPLE_NOTE_$0",
    SIMPLE_NOTE_$1 = "SIMPLE_NOTE_$1",
    NOTE_GROUP = "NOTE_GROUP",
    GROUP_START = "GROUP_START",
    GROUP_END = "GROUP_END",
    DURATION = "DURATION",
    DURATION_$0 = "DURATION_$0",
    DURATION_VALUE_1 = "DURATION_VALUE_1",
    DURATION_VALUE_2 = "DURATION_VALUE_2",
    DURATION_VALUE_3 = "DURATION_VALUE_3",
    NOTE_VALUE = "NOTE_VALUE",
    DURATION_SIGN = "DURATION_SIGN",
    SILENCE_SIGN = "SILENCE_SIGN",
    NOTE_SEPARATOR = "NOTE_SEPARATOR",
}
export interface BLOCK {
    kind: ASTKinds.BLOCK;
    head: BLOCK_CONTENT;
    tail: BLOCK_$0[];
}
export interface BLOCK_$0 {
    kind: ASTKinds.BLOCK_$0;
    content: BLOCK;
}
export type BLOCK_CONTENT = BLOCK_CONTENT_1 | BLOCK_CONTENT_2;
export interface BLOCK_CONTENT_1 {
    kind: ASTKinds.BLOCK_CONTENT_1;
    noteGroup: BLOCK_CONTENT_$0;
}
export interface BLOCK_CONTENT_2 {
    kind: ASTKinds.BLOCK_CONTENT_2;
    note: BLOCK_CONTENT_$1;
}
export type BLOCK_CONTENT_$0 = NOTE_GROUP;
export type BLOCK_CONTENT_$1 = NOTE;
export interface NOTE {
    kind: ASTKinds.NOTE;
    duration: Nullable<NOTE_$0>;
    simpleNote: SIMPLE_NOTE;
}
export type NOTE_$0 = DURATION;
export type SIMPLE_NOTE = SIMPLE_NOTE_1 | SIMPLE_NOTE_2;
export interface SIMPLE_NOTE_1 {
    kind: ASTKinds.SIMPLE_NOTE_1;
    silence: SIMPLE_NOTE_$0;
}
export interface SIMPLE_NOTE_2 {
    kind: ASTKinds.SIMPLE_NOTE_2;
    note: SIMPLE_NOTE_$1;
}
export type SIMPLE_NOTE_$0 = SILENCE_SIGN;
export type SIMPLE_NOTE_$1 = NOTE_VALUE;
export interface NOTE_GROUP {
    kind: ASTKinds.NOTE_GROUP;
    duration: DURATION;
    block: BLOCK;
}
export type GROUP_START = string;
export type GROUP_END = string;
export interface DURATION {
    kind: ASTKinds.DURATION;
    value: DURATION_$0;
}
export type DURATION_$0 = DURATION_VALUE;
export type DURATION_VALUE = DURATION_VALUE_1 | DURATION_VALUE_2 | DURATION_VALUE_3;
export type DURATION_VALUE_1 = string;
export type DURATION_VALUE_2 = string;
export type DURATION_VALUE_3 = string;
export type NOTE_VALUE = string;
export type DURATION_SIGN = string;
export type SILENCE_SIGN = string;
export type NOTE_SEPARATOR = [string, ...string[]];